API interceptor call for index
